creating database migrations and model relationship

+++ database/migrations/2023_08_15_084011_create_payment_table.php
@@ -16,7 +16,7 @@ class CreatePaymentTable extends Migration
         Schema::create('payment', function (Blueprint $table) {
             $table->id();
             $table->integer('payment_type');
-            $table->integer('total_price');
+            $table->decimal('total_price',10,2);
             $table->integer('order_id');
             $table->integer('transaction_code');
             $table->timestamps();
